+++ src/index.css
@@ -1,0 +1,13 @@
+html, body, #root {
+	height: 100%;
+}
+
+.row {
+	width: 100%;
+	height: 33%;
+}
+
+.container, .col {
+	height: 100%;
+	width: 100%;
+}
