+++ src/sockets.js
@@ -3,41 +3,41 @@ var socket = openSocket("http://localhost:8000");
 
 // subscribe
 
-function subscribeToGameStart(callback) {
+export function subscribeToGameStart(callback) {
 	socket.on("start", callback);
 }
 
-function subscribeToPositions(callback) {
+export function subscribeToPositions(callback) {
 	socket.on("positions", (otherPlayers, playerPosition) => {
 		callback(otherPlayers, playerPosition);
 	});
 }
 
-function subscribeToJoin(callback) {
+export function subscribeToJoin(callback) {
 	socket.on("join", (pos) => {
 		callback(pos);
 	});
 }
 
-function subscribeToLeave(callback) {
+export function subscribeToLeave(callback) {
 	socket.on("leave", (pos) => {
 		callback(pos);
 	});
 }
 
-function subscribeToCodeWords(callback) {
+export function subscribeToCodeWords(callback) {
 	socket.on("codeWords", (codeWords, playerPosition) => {
 		callback(codeWords, playerPosition);
 	});
 }
 
-function subscribeToShuffledDeck(callback) {
+export function subscribeToShuffledDeck(callback) {
 	socket.on("shuffledDeck", (shuffledDeck, playerPosition) => {
 		callback(shuffledDeck, playerPosition);
 	});
 }
 
-function subscribeToLockedDeck(callback) {
+export function subscribeToLockedDeck(callback) {
 	socket.on("lockedDeck", (lockedDeck, playerPosition) => {
 		callback(lockedDeck, playerPosition);
 	});
@@ -45,16 +45,14 @@ function subscribeToLockedDeck(callback) {
 
 // send
 
-function sendCodeWords(codeWords) {
+export function sendCodeWords(codeWords) {
 	socket.emit("codeWords", codeWords);
 }
 
-function sendShuffledDeck(shuffledDeck) {
+export function sendShuffledDeck(shuffledDeck) {
 	socket.emit("shuffledDeck", shuffledDeck);
 }
 
-function sendLockedDeck(lockedDeck) {
+export function sendLockedDeck(lockedDeck) {
 	socket.emit("lockedDeck", lockedDeck);
 }
-
-export { subscribeToGameStart, subscribeToPositions, subscribeToJoin, subscribeToLeave, subscribeToCodeWords, sendCodeWords, subscribeToShuffledDeck, sendShuffledDeck, subscribeToLockedDeck, sendLockedDeck };
